Add files via upload
Calculates the center of mass of the k rat by estimating the center of the distance between the eye and hip. A paired t test is also performed to see if there are statistical differences of the pitch angle and displacement while hopping on a sand versus solid surface. Violin plots and graphs of the COM displacement and pitch angle are also included.