Add files via upload
Interpolates the kinematic data from 300 to 400 Hz. This was necessary because to align the analog data with the Boris and kinematic data, the conversion sampling factor needs to be an integer (4000 Hz analog sampling rate and 400 Hz video sampling rate = a conversion factor of 10 samples of analog data per one kinematic point).